+++ package.json
@@ -10,7 +10,7 @@
   "engines": {
     "node": ">=6.0.0"
   },
-  "main": "index.js",
+  "main": "src/index.js",
   "stripes": {
     "type": "app",
     "displayName": "ui-checkout.meta.title",
