+++ src/components/ScanFooter/ScanFooter.css
