+++ src/components/ScanFooter/ScanFooter.js
