+++ src/components/ScanFooter/index.js
