+++ src/components/ScanTotal/ScanTotal.css
