+++ src/components/ScanTotal/ScanTotal.js
