+++ src/components/ScanTotal/index.js
